examples and docs
