+++ machinelearning/googlemachinelearning/googlemachinelearning_test.go
@@ -31,7 +31,7 @@ func TestUpdateMLModel(t *testing.T) {
 }
 
 
-
+/*
 func TestCreateMLModel(t *testing.T) {
 
 	var googlemachinelearning Googlemachinelearning
@@ -95,3 +95,4 @@ func TestDeleteMLModel(t *testing.T) {
 	response := resp.(map[string]interface{})
 	fmt.Println(response["body"])
 }
+*/
